+++ src/Figment.Common/Data/SchemaStorageProviderBase.cs
@@ -16,10 +16,6 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using System.Text;
-using System.Text.Json;
-using Figment.Common.Errors;
-
 namespace Figment.Common.Data;
 
 /// <summary>
@@ -27,18 +23,6 @@ namespace Figment.Common.Data;
 /// </summary>
 public abstract class SchemaStorageProviderBase : ISchemaStorageProvider
 {
-    /// <summary>
-    /// Gets the Json serialization options to use when serializing and deserializing content.
-    /// </summary>
-    protected static JsonSerializerOptions SerializerOptions { get; } = new()
-    {
-        // Required for $ref properties with type descriminator
-        AllowOutOfOrderMetadataProperties = true,
-#if DEBUG
-        WriteIndented = true,
-#endif
-    };
-
     /// <inheritdoc/>
     public abstract Task<CreateSchemaResult> CreateAsync(string schemaName, CancellationToken cancellationToken);
 
@@ -65,42 +49,6 @@ public abstract class SchemaStorageProviderBase : ISchemaStorageProvider
 
     /// <inheritdoc/>
     public abstract Task<Schema?> LoadAsync(string schemaGuid, CancellationToken cancellationToken);
-
-    /// <summary>
-    /// Loads schema from a serialized Json string.
-    /// </summary>
-    /// <param name="content">Json string to deserialize into a <see cref="Schema"/>.</param>
-    /// <param name="cancellationToken">The cancellation token.</param>
-    /// <returns>The deserliazed schema, or null if there was a serialization error.</returns>
-    public async Task<Schema?> LoadJsonContentAsync(string content, CancellationToken cancellationToken)
-    {
-        await using var ms = new MemoryStream(Encoding.UTF8.GetBytes(content));
-
-        try
-        {
-            var schemaDefinition = await JsonSerializer.DeserializeAsync<JsonSchemaDefinition>(ms, SerializerOptions, cancellationToken);
-            if (schemaDefinition == null)
-            {
-                AmbientErrorContext.Provider.LogError("Unable to deserialize schema from content");
-                return null;
-            }
-
-            var schema = schemaDefinition.ToSchema();
-
-            // Set name fields.
-            foreach (var prop in schema.Properties)
-            {
-                prop.Value.Name = prop.Key;
-            }
-
-            return schema;
-        }
-        catch (JsonException je)
-        {
-            AmbientErrorContext.Provider.LogException(je, "Unable to deserialize schema from content string");
-            return null;
-        }
-    }
 
     /// <inheritdoc/>
     public abstract Task<bool> RebuildIndexes(CancellationToken cancellationToken);